Move object initializer mutator from InitializerMutator to ObjectCreationMutator.
