+++ src/Stryker.Core/Stryker.Core/Mutators/ObjectCreationMutator.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Stryker.Core.Mutants;
@@ -20,6 +20,16 @@ namespace Stryker.Core.Mutators
                     ReplacementNode = node.ReplaceNode(node.Initializer, SyntaxFactory.InitializerExpression(SyntaxKind.CollectionInitializerExpression)),
                     DisplayName = "Collection initializer mutation",
                     Type = Mutator.Initializer
+                };
+            }
+            if (node.Initializer?.Kind() == SyntaxKind.ObjectInitializerExpression && node.Initializer.Expressions.Count > 0)
+            {
+                yield return new Mutation()
+                {
+                    OriginalNode = node,
+                    ReplacementNode = node.ReplaceNode(node.Initializer, SyntaxFactory.InitializerExpression(SyntaxKind.ObjectInitializerExpression)),
+                    DisplayName = "Object initializer mutation",
+                    Type = Mutator.Initializer,
                 };
             }
         }
